Cleaning up test structure a little
